refactor(llm): Move the shared chat-completions wire types to their own module
`cerebras` and `llamacpp` both speak the OpenAI `/v1/chat/completions`
dialect, and both need the same chunk types and assistant-message
merging to do it. Those lived in `llamacpp`, so `cerebras` reached into
a sibling provider for them and the shared thing carried the name of one
of its two users.

They now live in `provider::openai_compat`, named for the protocol
instead: the SSE chunk shapes, their `reasoning_content` extension, and
`merge_consecutive_assistant_messages`. Both providers import from it
and neither imports from the other. The module doc says what the dialect
is and how it differs from the Responses API that the `openai` provider
speaks, so the next contributor adding an OpenAI-compatible provider has
somewhere obvious to look.

Behavior is unchanged; this is a move plus the import rewiring it
forces. `parameters_with_strict_mode` is the remaining shared helper
still living in `openai`, used by four providers including the two here.
Leaving it alone: it is reached from a module that genuinely is about
OpenAI, and moving it would drag its test module along for no gain here.
